Move pillars implementation next to behaviour definition
